requirements.txt
Django==1.11.2
django-cors-headers==2.2.0
django-latency==0.0.1.1
djangorestframework==3.7.7
pytz==2018.3
